[ADD] Move aero and hydro elements to new fluid directory
