+++ include/seahowl/fluid/aero/bemt.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "seahowl/aero/blade_aero.h"
-#include "seahowl/aero/tower_aero.h"
+#include "seahowl/fluid/aero/blade_aero.h"
+#include "seahowl/fluid/aero/tower_aero.h"
 #include "seahowl/commons/numerics.h"
 
 // forward declarations
