+++ include/seahowl/fluid/aero/turbine_aero.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "seahowl/aero/rotor_aero.h"
-#include "seahowl/aero/tower_aero.h"
-#include "seahowl/hydro/foundation_fluid.h"
-#include "seahowl/hydro/floater_hydro.h"
+#include "seahowl/fluid/aero/rotor_aero.h"
+#include "seahowl/fluid/aero/tower_aero.h"
+#include "seahowl/fluid/hydro/foundation_fluid.h"
+#include "seahowl/fluid/hydro/floater_hydro.h"
 #include "seahowl/commons/component_fluid.h"
 
 #include <vector>
