+++ include/seahowl/fluid/hydro/mooring_hydro.h
@@ -2,7 +2,7 @@
 
 #include "seahowl/commons/numerics.h"
 #include "seahowl/commons/entities.h"
-#include "seahowl/hydro/morison.h"
+#include "seahowl/fluid/hydro/morison.h"
 #include "seahowl/commons/component_fluid.h"
 
 #include <vector>
