+++ src/bindings/python/pyseahowl_core.cpp
@@ -6,24 +6,24 @@
 #include <seahowl/core/simulation.h>
 #include <seahowl/core/turbine.h>
 #include <seahowl/elasto/turbine_elasto.h>
-#include <seahowl/aero/turbine_aero.h>
+#include <seahowl/fluid/aero/turbine_aero.h>
 #include <seahowl/core/rotor.h>
 #include <seahowl/core/blade.h>
 #include <seahowl/elasto/blade_elasto.h>
-#include <seahowl/aero/blade_aero.h>
+#include <seahowl/fluid/aero/blade_aero.h>
 #include <seahowl/core/tower.h>
 #include <seahowl/elasto/tower_elasto.h>
 #include <seahowl/core/mooring.h>
 #include <seahowl/elasto/mooring_elasto.h>
-#include <seahowl/hydro/mooring_hydro.h>
+#include <seahowl/fluid/hydro/mooring_hydro.h>
 #include <seahowl/core/foundation.h>
 #include <seahowl/core/floater.h>
 #include <seahowl/core/monopile.h>
 #include <seahowl/elasto/floater_elasto.h>
-#include <seahowl/hydro/floater_hydro.h>
+#include <seahowl/fluid/hydro/floater_hydro.h>
 #include <seahowl/core/system.h>
 #include <seahowl/elasto/system_elasto.h>
-#include <seahowl/aero/system_aero.h>
+#include <seahowl/fluid/aero/system_aero.h>
 #include <seahowl/env/wind_models.h>
 #include <seahowl/env/env_model.h>
 #include <seahowl/servo/controller.h>
