+++ src/bindings/python/pyseahowl_io.cpp
@@ -12,9 +12,9 @@
 #include <seahowl/elasto/blade_elasto.h>
 #include <seahowl/elasto/tower_elasto.h>
 #include <seahowl/elasto/rotor_elasto.h>
-#include <seahowl/aero/blade_aero.h>
-#include <seahowl/aero/tower_aero.h>
-#include <seahowl/aero/rotor_aero.h>
+#include <seahowl/fluid/aero/blade_aero.h>
+#include <seahowl/fluid/aero/tower_aero.h>
+#include <seahowl/fluid/aero/rotor_aero.h>
 #include <seahowl/io/write_csv.h>
 #include <seahowl/io/output_manager.h>
