+++ src/fluid/aero/disk.cpp
@@ -1,8 +1,8 @@
-#include <seahowl/aero/bemt.h>
+#include <seahowl/fluid/aero/bemt.h>
 
-#include <seahowl/aero/airfoil.h>
-#include <seahowl/aero/blade_aero.h>
-#include <seahowl/aero/tower_aero.h>
+#include <seahowl/fluid/aero/airfoil.h>
+#include <seahowl/fluid/aero/blade_aero.h>
+#include <seahowl/fluid/aero/tower_aero.h>
 
 #include <iostream>
