+++ src/fluid/aero/rotor_aero.cpp
@@ -1,9 +1,9 @@
-#include "seahowl/aero/rotor_aero.h"
+#include "seahowl/fluid/aero/rotor_aero.h"
 
-#include "seahowl/aero/blade_aero.h"
-#include "seahowl/aero/tower_aero.h"
-#include "seahowl/aero/airfoil.h"
-#include "seahowl/aero/bemt.h"
+#include "seahowl/fluid/aero/blade_aero.h"
+#include "seahowl/fluid/aero/tower_aero.h"
+#include "seahowl/fluid/aero/airfoil.h"
+#include "seahowl/fluid/aero/bemt.h"
 #include "seahowl/commons/utils.h"
 
 #include <cmath>
