+++ tests/unit_tests/test_morison.cpp
@@ -1,10 +1,10 @@
 #include "fixture_components.h"
 
-#include <seahowl/hydro/morison.h>
+#include <seahowl/fluid/hydro/morison.h>
 #include <seahowl/env/wave_models.h>
 #include <seahowl/core/simulation.h>
 #ifdef HAVE_HYDROCHRONO
-    #include <seahowl/hydro/hydrochrono_adapter.h>
+    #include <seahowl/fluid/hydro/hydrochrono_adapter.h>
     #include <hydroc/hydro_forces.h>
 #endif
