+++ tests/unit_tests/test_turbine.cpp
@@ -7,11 +7,11 @@
 #include <seahowl/env/wind_models.h>
 #include <seahowl/env/env_model.h>
 #include <seahowl/commons/numerics.h>
-#include <seahowl/aero/turbine_aero.h>
+#include <seahowl/fluid/aero/turbine_aero.h>
 #include <seahowl/core/turbine.h>
 #include <seahowl/core/system.h>
 #include <seahowl/servo/controller.h>
-#include <seahowl/aero/system_aero.h>
+#include <seahowl/fluid/aero/system_aero.h>
 #include <seahowl/io/read_input.h>
 
 #include <filesystem>  // C++17
